summary dependencies and add unit test
