+++ shell/src/main/java/alluxio/cli/fsadmin/command/ReportCommand.java
@@ -12,12 +12,16 @@
 package alluxio.cli.fsadmin.command;
 
 import alluxio.cli.AbstractCommand;
-import alluxio.cli.fsadmin.command.report.SummaryCommand;
+import alluxio.cli.fsadmin.report.SummaryCommand;
+import alluxio.client.MetaMasterClient;
+import alluxio.client.RetryHandlingMetaMasterClient;
+import alluxio.client.block.RetryHandlingBlockMasterClient;
 import alluxio.client.file.FileSystemContext;
 import alluxio.client.file.FileSystemMasterClient;
 import alluxio.exception.ExceptionMessage;
 import alluxio.exception.status.InvalidArgumentException;
 import alluxio.exception.status.UnavailableException;
+import alluxio.master.MasterClientConfig;
 import alluxio.master.MasterInquireClient;
 import alluxio.master.PollingMasterInquireClient;
 import alluxio.resource.CloseableResource;
@@ -37,6 +41,8 @@ import javax.annotation.concurrent.ThreadSafe;
  */
 @ThreadSafe
 public final class ReportCommand extends AbstractCommand {
+  private MetaMasterClient mMetaMasterClient;
+  private RetryHandlingBlockMasterClient mBlockMasterClient;
 
   enum Command {
     SUMMARY // Reports the Alluxio cluster information
@@ -45,7 +51,10 @@ public final class ReportCommand extends AbstractCommand {
   /**
    * Creates a new instance of {@link ReportCommand}.
    */
-  public ReportCommand() {}
+  public ReportCommand() {
+    mMetaMasterClient = new RetryHandlingMetaMasterClient(MasterClientConfig.defaults());
+    mBlockMasterClient = new RetryHandlingBlockMasterClient(MasterClientConfig.defaults());
+  }
 
   @Override
   public String getCommandName() {
@@ -59,56 +68,64 @@ public final class ReportCommand extends AbstractCommand {
 
   @Override
   public int run(CommandLine cl) throws IOException {
-    // Checks if Alluxio master and client services are running
-    try (CloseableResource<FileSystemMasterClient> client =
-             FileSystemContext.INSTANCE.acquireMasterClientResource()) {
-      MasterInquireClient inquireClient = null;
-      try {
-        InetSocketAddress address = client.get().getAddress();
-        List<InetSocketAddress> addresses = Arrays.asList(address);
-        inquireClient = new PollingMasterInquireClient(addresses, () ->
-            new ExponentialBackoffRetry(50, 100, 2));
-      } catch (UnavailableException e) {
-        System.err.println("Failed to get the leader master.");
-        System.err.println("Please check your Alluxio master status");
-        return 1;
-      }
-      try {
-        inquireClient.getPrimaryRpcAddress();
-      } catch (UnavailableException e) {
-        System.err.println("The Alluxio leader master is not currently serving requests.");
-        System.err.println("Please check your Alluxio master status");
-        return 1;
-      }
-    }
-
-    String[] args = cl.getArgs();
-
-    if (args.length == 0) {
-      SummaryCommand summaryCommand = new SummaryCommand();
-      summaryCommand.run();
-      return 0;
-    }
-
-    // Gets the report category
-    Command command;
     try {
-      String commandName = args[0].toUpperCase();
-      command = Command.valueOf(commandName);
-    } catch (IllegalArgumentException e) {
-      System.out.println(getUsage());
-      System.out.println(getDescription());
-      return -1;
-    }
+      // Checks if Alluxio master and client services are running
+      try (CloseableResource<FileSystemMasterClient> client =
+               FileSystemContext.INSTANCE.acquireMasterClientResource()) {
+        MasterInquireClient inquireClient = null;
+        try {
+          InetSocketAddress address = client.get().getAddress();
+          List<InetSocketAddress> addresses = Arrays.asList(address);
+          inquireClient = new PollingMasterInquireClient(addresses, () ->
+              new ExponentialBackoffRetry(50, 100, 2));
+        } catch (UnavailableException e) {
+          System.err.println("Failed to get the leader master.");
+          System.err.println("Please check your Alluxio master status");
+          return 1;
+        }
+        try {
+          inquireClient.getPrimaryRpcAddress();
+        } catch (UnavailableException e) {
+          System.err.println("The Alluxio leader master is not currently serving requests.");
+          System.err.println("Please check your Alluxio master status");
+          return 1;
+        }
+      }
 
-    switch (command) {
-      case SUMMARY:
-        SummaryCommand summaryCommand = new SummaryCommand();
+      String[] args = cl.getArgs();
+
+      // Prints the summarized information in default situation
+      if (args.length == 0) {
+        SummaryCommand summaryCommand = new SummaryCommand(mMetaMasterClient,
+            mBlockMasterClient);
         summaryCommand.run();
-        break;
-      // CAPACITY, CONFIGURATION, RPC, OPERATION, and UFS commands will be supported in the future
-      default:
-        break;
+        return 0;
+      }
+
+      // Gets the report category
+      Command command;
+      try {
+        String commandName = args[0].toUpperCase();
+        command = Command.valueOf(commandName);
+      } catch (IllegalArgumentException e) {
+        System.out.println(getUsage());
+        System.out.println(getDescription());
+        return -1;
+      }
+
+      switch (command) {
+        case SUMMARY:
+          SummaryCommand summaryCommand = new SummaryCommand(mMetaMasterClient,
+              mBlockMasterClient);
+          summaryCommand.run();
+          break;
+        // CAPACITY, CONFIGURATION, RPC, OPERATION, and UFS commands will be supported in the future
+        default:
+          break;
+      }
+    } finally {
+      mMetaMasterClient.close();
+      mBlockMasterClient.close();
     }
     return 0;
   }
